cli clean version
