+++ src/cli_app.py
@@ -1,40 +1,31 @@
-from preprocessing import get_dataset_from_path, preprocess_data
-from prediction import predict, tokenizer, model, label2id
+from preprocessing import get_dataset_from_path, add_token_ids, preprocess_data, label2id
+from prediction import predict, tokenizer
 from postprocessing import postprocess_data
-import numpy as np
-import pandas as pd
 
-# predict_pii.py: cmd version, takes a filepath with a batch of essays as an argument, does the preprocessing,
-# predictions, postprocessing and generates a csv
+# cli_app.py: cmd version, takes a filepath with a batch of essays as an argument, does the preprocessing,
+# the predictions, the postprocessing and generates a csv with all the retrieved PII
 
 
 if __name__ == "__main__":
 
     help = """
     Welcome to the Personally Identifiable Information detection system.\n
-    0.5 Download the fine-tuned model (this will be done automatically)
-    1. Type the path to a batch of student essays that contain some PII in .json or .csv format.
-    2. The system will pass the essays through a fine-tuned LLM and assign a label to every word.
+    1. Type the path to a batch of student essays that contain some PII in .json or .csv format (usable example: data/test/ten_essays_test.json).
+    2. The system will pass the essays through a fine-tuned LLM (RoBERTa) and label every token.
     3. The system will save all the tokens that are some kind of PII in the result_pii.csv file.
-    The columns in the output file will correspond to the essay id, the token string and the assigned (non-O) label.
+    The columns in the output file will correspond to the document (essay) id, the token id in the essay, the token string and the assigned (non-O) label.
     """
     print(help)
 
-    data_path = input("Input your path to the .json datafile: ")
+    data_path = input(
+        "Input your path to the .json datafile (data/test/ten_essays_test.json or other): ")
     dataset = get_dataset_from_path(data_path)
     dataset = dataset.rename_column("tokens", "token_string")
-
-    def add_token_ids(example):
-        return {'token_id': [i for i in range(len(example['token_string']))]}
-    data = dataset.map(add_token_ids)
+    dataset = dataset.map(add_token_ids)
     data_tokenized = preprocess_data(
-        data, tokenizer, label2id=label2id, with_labels=False)
+        dataset, tokenizer, label2id=label2id, with_labels=False)
     predictions = predict(data_tokenized)
-    final_df = postprocess_data(data_tokenized, predictions)
-    ds_df = data.to_pandas()
-    ds_df = ds_df.explode(['token_id', 'token_string'])
-    df_back_to_token_string = pd.merge(
-        ds_df, final_df, how='inner', on=['token_id'])
-    aligned_pii = df_back_to_token_string[['token_string', 'label']]
-    print(aligned_pii)
+    result_pii = postprocess_data(
+        dataset, data_tokenized, predictions)
+    result_pii.to_csv("result_pii.csv", index=False)
     print(f"Successfully saved predictions for your file in result_pii.csv")
