+++ src/prediction.py
@@ -6,23 +6,6 @@ tokenizer = AutoTokenizer.from_pretrained(
     'roberta-base', add_prefix_space=True)
 model = AutoModelForTokenClassification.from_pretrained(
     "zeinab-sheikhi/roberta-pii-detection")
-label2id = {
-    'B-NAME_STUDENT': 0,
-    'B-EMAIL': 1,
-    'B-USERNAME': 2,
-    'B-ID_NUM': 3,
-    'B-PHONE_NUM': 4,
-    'B-URL_PERSONAL': 5,
-    'B-STREET_ADDRESS': 6,
-    'I-NAME_STUDENT': 7,
-    'I-EMAIL': 8,
-    'I-USERNAME': 9,
-    'I-ID_NUM': 10,
-    'I-PHONE_NUM': 11,
-    'I-URL_PERSONAL': 12,
-    'I-STREET_ADDRESS': 13,
-    'O': 14,
-    '[PAD]': -100}
 
 
 def get_embeddings(example):
@@ -37,6 +20,4 @@ def predict(preprocessed_data):
     logits = torch.nn.functional.softmax(outputs['logits'], dim=-1)
     logits = (outputs['logits']).detach().numpy()
     predictions = np.argmax(logits, axis=-1)
-    print(outputs)
-    print(len(predictions))
     return predictions
